Topology compaction omits image records if minEpoch is ahead of last image Also Fix:  - RegisteredCallback should remove itself from callback map when cancelled  - Do not throw CancellationException when processing requests that have been aborted, as may be caused by a successful meaningful reply that can be overridden  - system_accord_debug.{executors,coordinations} fail with ClassCastException  - CommandStore.updateMinHlc eats up CPU as called much too often  - AccordCache not notifying flushed on shutdown Also Improve:  - Support skipping Deps  - Violation information reported  - Sort CommandStore shards by id
patch by Benedict; reviewed by Aleksey Yeschenko for CASSANDRA-20896

+++ src/java/org/apache/cassandra/service/accord/serializers/SerializePacked.java
@@ -133,13 +133,13 @@ public class SerializePacked
         }
     }
 
-    public static long serializedPackedSize(int count, long max)
+    public static int serializedPackedSize(int count, long max)
     {
         return serializedPackedBitsSize(count, BitUtils.numberOfBitsToRepresent(max));
     }
 
-    public static long serializedPackedBitsSize(int count, int bitsPerEntry)
+    public static int serializedPackedBitsSize(int count, int bitsPerEntry)
     {
-        return ((long)bitsPerEntry * count + 7)/8;
+        return (int) (((long)bitsPerEntry * count + 7) / 8);
     }
 }
